pyproject.toml
[project]
name = "blog-toolkit"
version = "0.1.0"
description = "A toolkit for collecting, storing, and analyzing multiple blogs with RSS feed and web crawler support"
readme = "README.md"
authors = [
    { name = "Leo Guinan", email = "leo.guinan@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "click>=8.1.0",
    "flask>=3.0.0",
    "pandas>=2.1.0",
    "nltk>=3.8.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "python-dateutil>=2.8.0",
    "textstat>=0.7.0",
    "vaderSentiment>=3.3.2",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
blog-toolkit = "blog_toolkit.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
